+++ GUI/Misc/AlertSystem.lua
@@ -5,8 +5,9 @@ gui.AlertSystem = {};
 local alertSystem = gui.AlertSystem;
 local helperFrame;
 
-local function ShowActiveEvents(showPopUps)
-    if not showPopUps then
+local function ShowActiveEvents()
+    local isInInstance = (select(3, GetInstanceInfo())) ~= 0;
+    if not addon.Options.db.EventReminders.ShowPopUps.OnLogin or (not addon.Options.db.EventReminders.ShowPopUps.OnLoginInInstances and isInInstance) then
         return;
     end
 
@@ -22,28 +23,32 @@ local function OnUpdate(self, elapsed)
     if self.TimeSinceLastUpdate > addon.Options.db.EventReminders.RefreshInterval then
         self.TimeSinceLastUpdate = 0;
         local currActiveEvents = addon.EventData:GetActiveEvents(true); -- This does the refresh of the event data and should be the only location
-        if addon.Options.db.EventReminders.ShowPopUps.OnEventStart then
-            prevActiveEvents = prevActiveEvents or currActiveEvents;
-            local matchFound;
-            for _, currActiveEvent in next, currActiveEvents do
-                for _, prevActiveEvent in next, prevActiveEvents do
-                    if currActiveEvent.Id == prevActiveEvent.Id then
-                        matchFound = true;
-                    end
-                end
-                if not matchFound then
-                    addon.Diagnostics.Print("New event, show pop up", currActiveEvent.Id, currActiveEvent.EventDetails.Name);
-                    alertSystem:AddAlert(currActiveEvent, addon.Options.db.EventReminders.FadeDelay);
+        local isInInstance = (select(3, GetInstanceInfo())) ~= 0;
+        if not addon.Options.db.EventReminders.ShowPopUps.OnEventStart or (not addon.Options.db.EventReminders.ShowPopUps.OnEventStartInInstances and isInInstance) then
+            return;
+        end
+
+        prevActiveEvents = prevActiveEvents or currActiveEvents;
+        local matchFound;
+        for _, currActiveEvent in next, currActiveEvents do
+            matchFound = false;
+            for _, prevActiveEvent in next, prevActiveEvents do
+                if currActiveEvent.Id == prevActiveEvent.Id then
+                    matchFound = true;
                 end
             end
-            prevActiveEvents = currActiveEvents
+            if not matchFound then
+                addon.Diagnostics.Print("New event, show pop up", currActiveEvent.Id, currActiveEvent.EventDetails.Name);
+                alertSystem:AddAlert(currActiveEvent, addon.Options.db.EventReminders.FadeDelay);
+            end
         end
+        prevActiveEvents = currActiveEvents;
     end
 end
 
 local function SetUp(frame, event, duration)
 	frame.Event = event;
-	frame.Name:SetText(event.EventDetails.Name);
+	frame.Name:SetText(event.EventDetails and event.EventDetails.Name or addon.L["Collecting data"]);
     frame:UpdateEventRuntime();
 	frame.Icon.Texture:SetTexture(event.Icon);
     frame.duration = duration;